% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densBB.R
\name{densBB}
\alias{densBB}
\title{Calculation of the beta-binomial joint density function}
\usage{
densBB(
  x1,
  x2,
  n1,
  n2,
  u,
  l,
  a0,
  a1,
  a2,
  i = 1,
  equal_thetas = FALSE,
  use_log = FALSE
)
}
\arguments{
\item{x1}{Numeric vector. Observations for the first sample.}

\item{x2}{Numeric vector. Observations for the second sample.}

\item{n1}{Integer. Number of trials in the first sample.}

\item{n2}{Integer. Number of trials in the second sample.}

\item{u}{Numeric. Upper argument for the generalized hypergeometric function.}

\item{l}{Numeric. Lower argument for the generalized hypergeometric function.}

\item{a0}{Numeric. First shape parameter of the beta-binomial distribution.}

\item{a1}{Numeric. Second shape parameter of the beta-binomial distribution.}

\item{a2}{Numeric. Third shape parameter of the beta-binomial distribution.}

\item{i}{Integer. Index for the observation to calculate in both samples. Default is 1.}

\item{equal_thetas}{Logical. If TRUE, theta1 and theta2 will be forced to be equal. Default is FALSE.}

\item{use_log}{Logical. If TRUE, the function returns the log of the density. Default is FALSE.}
}
\value{
A function that computes the beta-binomial joint density for the given parameters.
}
\description{
This function returns a functor that calculates the joint density of two
random variables following a beta-binomial distribution.
}
\examples{
dens_fun <- densBB_functor(x1 = c(0, 1), x2 = c(1, 1), n1 = 10, n2 = 12,
                           u = 1, l = 1, a0 = 0.5, a1 = 1, a2 = 1, i = 1)
dens_fun(0.6, 0.7)

}
