% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_two.R
\name{error_two}
\alias{error_two}
\title{Calculate Error Based on Evidence}
\usage{
error_two(
  x1,
  x2,
  n1,
  n2,
  alpha0,
  alpha1,
  alpha2,
  i,
  j,
  N2,
  n_bootstrap,
  k,
  seed = NULL
)
}
\arguments{
\item{x1}{Numeric vector. Observations for the first sample.}

\item{x2}{Numeric vector. Observations for the second sample.}

\item{n1}{Integer. Number of trials for the first sample.}

\item{n2}{Integer. Number of trials for the second sample.}

\item{alpha0}{Numeric. First shape parameter.}

\item{alpha1}{Numeric. Second shape parameter.}

\item{alpha2}{Numeric. Third shape parameter.}

\item{i}{Integer. Index for the observation in both samples.}

\item{j}{Integer. Index for Theta generation.}

\item{N2}{Integer. Number of samples for `Theta1`.}

\item{n_bootstrap}{Integer. Number of bootstrap samples for variational Bayes.}

\item{k}{Numeric. Threshold for calculating the indicator `Ind`.}

\item{seed}{Integer. Seed for random number generation. Default is `NULL`, which generates a random seed.}
}
\value{
Numeric. The calculated error `error2`.
}
\description{
This function calculates an error metric (`error2`) based on the evidence calculated by the `evidence` function under type II error.
It includes checks for `NA` values in the calculated evidence (`ev10`) and verification value (`f_ver`),
and throws an error if any of them are `NA`.
}
\examples{
error_one(x1 = c(1, 2), x2 = c(2, 3), n1 = 10, n2 = 12, alpha0 = 1, alpha1 = 2, alpha2 = 3, i = 1, j = 1, N2 = 100, n_bootstrap = 1000, k = 0.5)

}
