% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_evidence.R
\name{evidence}
\alias{evidence}
\title{Calculate Evidence for the Beta-Binomial Model}
\usage{
evidence(
  x1,
  x2,
  n1,
  n2,
  alpha0,
  alpha1,
  alpha2,
  i,
  j,
  N2,
  n_bootstrap,
  seed = NULL
)
}
\arguments{
\item{x1}{Numeric vector. Observations for the first sample.}

\item{x2}{Numeric vector. Observations for the second sample.}

\item{n1}{Integer. Number of trials for the first sample.}

\item{n2}{Integer. Number of trials for the second sample.}

\item{alpha0}{Numeric. First shape parameter.}

\item{alpha1}{Numeric. Second shape parameter.}

\item{alpha2}{Numeric. Third shape parameter.}

\item{i}{Integer. Index for the observation in both samples.}

\item{j}{Integer. Index for Theta sampling.}

\item{N2}{Integer. Number of samples for `Theta1`.}

\item{n_bootstrap}{Integer. Number of bootstrap samples for variational Bayes.}

\item{seed}{Integer. Seed for random number generation. Default is `NULL`, which generates a random seed.}
}
\value{
Numeric. The calculated evidence `ev10`.
}
\description{
This function calculates the evidence for a beta-binomial model using variational inference under error type I
via Stan and genetic algorithms to optimize parameters. It returns the calculated evidence (`ev10`).
}
\examples{
evidence(x1 = c(1, 0), x2 = c(1, 1), n1 = 2, n2 = 2, alpha0 = 0.6, alpha1 = 0.65, alpha2 = 0.7, i = 1, j = 1, N2 = 100, n_bootstrap = 1000)

}
