% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_value_beta.R
\name{full_evidence_beta}
\alias{full_evidence_beta}
\title{Calculate Evidence for Multiple Grids of Alpha}
\usage{
full_evidence_beta(
  x1,
  x2,
  n1,
  n2,
  alpha0,
  alpha1,
  alpha2,
  N1,
  N2,
  n_bootstrap,
  seed = NULL
)
}
\arguments{
\item{x1}{Numeric vector. Observations for the first sample.}

\item{x2}{Numeric vector. Observations for the second sample.}

\item{n1}{Integer. Number of trials for the first sample.}

\item{n2}{Integer. Number of trials for the second sample.}

\item{alpha0}{Numeric. First shape parameter.}

\item{alpha1}{Numeric. Second shape parameter.}

\item{alpha2}{Numeric. Third shape parameter.}

\item{N1}{Integer. Number of iterations for the first loop.}

\item{N2}{Integer. Number of iterations for the second loop.}

\item{n_bootstrap}{Integer. Number of bootstrap samples.}

\item{seed}{Integer. Seed for random number generation. Default is `NULL`, which generates a random seed.}
}
\value{
A list where each element corresponds to the results of evidence calculations for a specific grid of `k`.
}
\description{
This function calculates the evidence (`beta_results_all_grids`) for multiple grids of values `k`
using parallel computation. It uses the `furrr` and `future` packages to distribute the computation
across multiple cores. The function calls `alf_integ2` for each value in the grid, passing the required parameters.
}
\examples{
full_evidence_alpha(x1 = c(1, 2), x2 = c(2, 3), n1 = 10, n2 = 12, alpha0 = 1, alpha1 = 2, alpha2 = 3, N1 = 100, N2 = 100, n_bootstrap = 1000)

}
